refactor: reorganize schema builder code
