+++ src/core/builders/index.ts
@@ -1,3 +1,3 @@
-export { FormFields } from "./form-fields";
-export { FormSchemaBuilder } from "./form-schema-builder";
+export { FormFields, FormSchemaBuilder } from "./schema";
+
 export { createFormValidator } from "./create-form-validator";
